requirements.txt
Flask==1.1.2
psutil==5.7.3
py-cpuinfo==7.0.0
matplotlib==3.3.3
python-dotenv==0.15.0